python/demo1_FormatFerry/requirements.txt
python-docx
deep-translator
opencc-python-reimplemented
